Renovation
- rename package
- rename command utility
- use six for compatibility
