+++ .gitignore
@@ -1,0 +1,4 @@
+.venv
+build
+dist
+venv
